feat: move javascript files to /sink
