+++ packages/substream/Dockerfile
@@ -17,7 +17,7 @@ FROM base
 COPY --from=build /app/dist /app/dist
 COPY --from=build /app/node_modules /app/node_modules
 COPY --from=build /app/package.json /app/package.json
-COPY --from=build /app/src/sql /app/src/sql
+COPY --from=build /app/src/sink /app/src/sink
 
 # TODO: Figure out how to set up schema for DB as part of
 # postgres setup
