+++ packages/substream/package.json
@@ -12,12 +12,12 @@
   "scripts": {
     "build": "tsc --noEmit && esbuild index.ts --bundle --outfile=dist/index.js --bundle --platform=node",
     "prepublishOnly": "npm run build",
-    "nuke-public": "tsx ./src/utils/executeSql.ts ./src/sql/nukePublic.sql",
+    "nuke-public": "tsx ./sink/utils/executeSql.ts ./sink/sql/nukePublic.sql",
     "zapatos": "pnpm dotenv -e .env -- npx zapatos",
     "start:from-genesis": "pnpm dotenv -e .env -- node dist/index.js --from-genesis",
     "start:from-genesis-with-cache": "pnpm dotenv -e .env -- node dist/index.js --from-cache --from-genesis",
     "start": "pnpm dotenv -e .env -- node dist/index.js",
-    "initialize": "./src/sql/initAll.sh && pnpm zapatos"
+    "initialize": "./sink/sql/initAll.sh && pnpm zapatos"
   },
   "dependencies": {
     "@bufbuild/buf": "^1.28.0",
