+++ packages/substream/sink/cursor.ts
