+++ packages/substream/sink/map-entries.ts
