+++ packages/substream/sink/parse-valid-full-entries.ts
