+++ packages/substream/sink/populate-entries.ts
