+++ packages/substream/sink/populate-roles.ts
