+++ packages/substream/sink/sql/clearAll.sh
@@ -3,7 +3,7 @@
 source .env
 
 echo "Running SQL scripts to clear all tables..."
-psql $DATABASE_URL < src/sql/clearCacheTables.sql
-psql $DATABASE_URL < src/sql/clearPublicTables.sql
+psql $DATABASE_URL < sink/sql/clearCacheTables.sql
+psql $DATABASE_URL < sink/sql/clearPublicTables.sql
 echo "SQL scripts executed successfully."
