+++ packages/substream/sink/sql/clearCacheTables.sh
@@ -3,6 +3,6 @@
 source .env
 
 echo "Running SQL scripts..."
-psql $DATABASE_URL < src/sql/clearCacheTables.sql
+psql $DATABASE_URL < sink/sql/clearCacheTables.sql
 echo "SQL scripts executed successfully."
