+++ packages/substream/sink/sql/clearPublicTables.sql
