+++ packages/substream/sink/sql/initIndexes.sql
