+++ packages/substream/sink/substreams.js/sink/src/index.ts
