+++ packages/substream/sink/substreams.js/sink/src/sink.ts
