+++ packages/substream/sink/substreams.js/sink/src/stream.ts
