+++ packages/substream/sink/utils/file-cursor.ts
