+++ packages/substream/sink/utils/id.ts
