+++ packages/substream/sink/zapatos/schema.d.ts
