+++ packages/substream/sink/zod.ts
