+++ packages/substream/tsconfig.json
@@ -39,5 +39,24 @@
     // "composite": true,
     // "declarationMap": true,
   },
-  "include": ["index.ts", "src/**/*", "schema.d.ts"]
+  "include": [
+    "index.ts",
+    "sink/**/*",
+    "schema.d.ts",
+    "sink/bootstrap-root.ts",
+    "sink/cursor.ts",
+    "sink/map-entries.ts",
+    "sink/parse-valid-full-entries.ts",
+    "sink/populate-entries.ts",
+    "sink/populate-from-cache.ts",
+    "sink/populate-profiles.ts",
+    "sink/populate-roles.ts",
+    "sink/run-stream.ts",
+    "sink/types.ts",
+    "sink/zod.ts",
+    "sink/zapatos",
+    "sink/utils",
+    "sink/substreams.js",
+    "sink/constants"
+  ]
 }
